:truck: Extract Sketch related symbols to its own package
