+++ packages/fyndiq-ui-sketch/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "fyndiq-ui-test",
+  "name": "fyndiq-ui-sketch",
   "private": true,
-  "version": "2.1.29",
+  "version": "2.0.0",
   "author": "Thibaut REMY <thibaut.remy@fyndiq.com>",
   "license": "MIT",
+  "main": "lib/index.js",
   "dependencies": {
     "fyndiq-component-alert": "^2.1.5",
     "fyndiq-component-article": "^2.0.0",
@@ -25,11 +26,6 @@
     "fyndiq-icons": "^2.8.0",
     "fyndiq-illustrations": "^2.1.1",
     "fyndiq-styles-colors": "^2.3.0",
-    "fyndiq-styles-fonts": "^2.1.0",
-    "fyndiq-ui-sketch": "*"
-  },
-  "scripts": {
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "fyndiq-styles-fonts": "^2.1.0"
   }
 }
