+++ packages/fyndiq-ui-sketch/src/index.js
@@ -1,0 +1,5 @@
+import symbols from './symbols'
+import SymbolComponent from './symbol-component'
+import Symbols from './symbols-component'
+
+export { symbols, SymbolComponent, Symbols }
