+++ packages/fyndiq-ui-sketch/src/symbols-component.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+import SymbolComponent from './symbol-component'
+
+export default ({ symbols }) =>
+  Object.keys(symbols).map(name => (
+    <SymbolComponent symbol={symbols[name]} name={name} key={name} />
+  ))
